requirements.txt
numpy==1.17.1
Pillow==6.1.0
PyAutoGUI==0.9.47
